Move quick_play to separate module.
To prevent circular import
